refactor(api): migrate article routes to new db layer and improve handlers
