+++ src/app/api/articles/route.ts
@@ -1,12 +1,19 @@
-import { connectToDatabase } from "@/lib/mongoose";
-import "@/models/Author";
-import "@/models/Tag";
-import Article from "@/models/Article";
+import { NextRequest, NextResponse } from "next/server";
+import { connectToDatabase } from "@/lib/db/mongoose";
+import "@/lib/models/Author";
+import "@/lib/models/Tag";
+import Article from "@/lib/models/Article";
 
-export async function GET() {
+export async function GET(request: NextRequest) {
   await connectToDatabase();
 
   try {
+    const { searchParams } = new URL(request.url);
+    const locale = searchParams.get("locale"); // "en", "es", "pt", etc. or null
+
+    // For now, we don't filter by locale yet.
+    // But we DO read it so the contract with useArticles is explicit and future-proof.
+
     const articles = await Article.find()
       .sort({
         createdAt: -1,
@@ -14,9 +21,12 @@ export async function GET() {
       .populate("author")
       .populate("tags");
 
-    return Response.json(articles);
+    return NextResponse.json(articles);
   } catch (error) {
     console.error("Error fetching articles:", error);
-    return new Response("Failed to fetch articles", { status: 500 });
+    return NextResponse.json(
+      { error: "Failed to fetch articles" },
+      { status: 500 }
+    );
   }
 }
