+++ src/app/api/test/route.ts
@@ -1,4 +1,4 @@
-import { connectToDatabase } from "@/lib/mongoose";
+import { connectToDatabase } from "@/lib/db/mongoose";
 import { NextResponse } from "next/server";
 
 export async function GET() {
